芝生の微調整とカラー設定のクラス作成
