+++ client/emil/emil/ColorSetting.swift
@@ -1,0 +1,39 @@
+//
+//  ColorSetting.swift
+//  emil
+//
+//  Created by 山川拓也 on 2017/11/08.
+//  Copyright © 2017年 !kie. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    class func lightBlue() -> UIColor {
+        return UIColor(red: 92.0 / 255, green: 192.0 / 255, blue: 210.0 / 255, alpha: 1.0)
+    }
+    
+    class func lightRed() -> UIColor {
+        return UIColor(red: 195.0 / 255, green: 123.0 / 255, blue: 175.0 / 255, alpha: 1.0)
+    }
+    
+    class func orange0() -> UIColor {
+        return UIColor(red: 254.0 / 255, green: 241.0 / 255, blue: 229.0 / 255, alpha: 1.0)
+    }
+    
+    class func orange1() -> UIColor {
+        return UIColor(red: 248.0 / 255, green: 198.0 / 255, blue: 149.0 / 255, alpha: 1.0)
+    }
+    
+    class func orange2() -> UIColor {
+        return UIColor(red: 240.0 / 255, green: 175.0 / 255, blue: 110.0 / 255, alpha: 1.0)
+    }
+    
+    class func orange3() -> UIColor {
+        return UIColor(red: 238.0 / 255, green: 150.0 / 255, blue: 64.0 / 255, alpha: 1.0)
+    }
+    
+    class func orange4() -> UIColor {
+        return UIColor(red: 229.0 / 255, green: 116.0 / 255, blue: 15.0 / 255, alpha: 1.0)
+    }    
+}
